Take an long chainID

+++ ethfirewall/src/main/java/tech/pegasys/ethfirewall/signing/ChainIdProvider.java
@@ -15,5 +15,5 @@ package tech.pegasys.ethfirewall.signing;
 /** Obtains the chain ID for the blockchain that transaction are being signed. */
 public interface ChainIdProvider {
 
-  byte id();
+  long id();
 }
